fabs -> fabsf

+++ common/fuse.h
@@ -221,7 +221,7 @@ public:
     //wGyro += (accel_ - down_).len2();
 
     // If acceleration is not 1.0G, don't trust it.
-    wGyro += fabs(accel_.len() - 1.0f) * 50.0;
+    wGyro += fabsf(accel_.len() - 1.0f) * 50.0;
 
     // If acceleration is changing rapidly, don't trust it.
     wGyro += accel_extrapolator_.slope().len() * 1000;
